tdf#136208 make combo/list/spin take focus on use of scroll wheel
So using the scrollwheel on those widgets when they do not have focus will
cause then to take focus. That will then cause the fontsize/fontname widgets to
reset back to the current document size/name if they are scrolled to a new
value, but their new value not explicitly applied to the current document,
when focus is restored to the document window.

Change-Id: I5d27e5e90c4d76d9cb3321878d82e22d40cacf81
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/102261
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>