sw: rename CppunitTest_sw_updateall_object_replacements to sw_uibase_uiview
All these new test suites are named in a way, so that in case the fix is
in sw/source/foo/bar/, then the matching test suite is sw_foo_bar.
Rename to this schema, so next time a bug is fixed in that directory, we
don't need to add a new suite.

Change-Id: I968711754cb587cc2f97fff6431be416b477728f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/102274
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>