Revert "tdf#59674 Word export: set a tabstop at numbering IndentAt"
This reverts commit 0409e651b72d275733aaf81f79814af2c6d41df6
for the benefit of all mankind, and specifically tdf#136644.

Reason for revert: LO's UI for this is totally messed up,
so I want to completely get out of anything to do with
numbering/bullets.

It is inconsistent with how it handles default tab stops
(delete all when none are shown and things act different),
it doesn't always increment levels properly,
and does all kinds of weird things.

Change-Id: Ifaf8c3a7c9ac30b7103b82dd119cb4629f0b2578
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/102415
Tested-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Justin Luth <justin_luth@sil.org>