tdf#136706 writerfilter: don't create page style for footnote
Each time CloseSectionGroup is called for a non-continuous
break, it wants to create a new page style.
Each footnote calls CloseSectionGroup twice,
so two unused ConvertedX page styles were created
for each footnote.

A quick glance through the code makes me wonder
whether footnotes can skip CloseSectionGroup
altogether, (unit tests suggest that it can...)
but it might be necessary for a floating table
that should be in-lined, or setting relative width
(although tables aren't really supported in footnotes).
In any case, the safest approach is to change as little
as possible, so that is what I have done.

HandleMargins...() only seems to add the margin information
into the page style, so it should be safe to avoid that
function, even though relative width stuff checks the margins.

Comments are already excluded, and the function is not even
called for isInIndexContext() || isInBibliographyContext().
I thought header/footers were also considered sections,
but that does not appear to be the case.
I'm not aware of any other types of non-text sections,
but if there are, they may need similar treatment.

Change-Id: I4325604c286ca1132e5765a56627be4b7e64ba4f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/102723
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>