tdf#136620 tdf#136708 filter,oox,sw: fix export of 2 different wraps
This reverts commit 2ea85df8af426eba8fa77e65b35d7f7d5a718b21.

Revert the change to EscherPropertyContainer, which was completely
bogus, based on pre-existing bogus code in VMLExport::Commit().

The problem is that ESCHER_Wrap values are for wrapping text *inside* a
text box, which is "mso-wrap-style" in VML, whereas VML's w10:wrap
element defines how text wraps *around* a shape, doesn't exist as an
ESCHER property and is specific to Word formats.

Instead, export the w10:wrap element in VMLExport::EndShape().

This has 2 callers, WriteActiveXControl() and writeVMLDrawing().

Furthermore the value "none" wasn't written for WrapTextMode_THROUGH,
which caused the wrap element to be omitted in that case.

Change-Id: Id4a01fcb2ea73fa9bef4ee8769b5e0680e059f15
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/103009
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>