tdf#136706 writerfilter: revert my info-gathering assert
The assert was failing in crashTesting for:
wget http://bugs.documentfoundation.org/attachment.cgi?id=68205
    -O tdf55725-1.docx
./instdir/program/soffice.bin --headless --convert-to docx
    ./tdf55725-1.docx

and also for
wget https://bugs.documentfoundation.org/attachment.cgi?id=127288
    -O tdf102131-1.docx

Both of these documents have SAX errors when loading.
Running these interactively, saying "no, do not continue loading"
parses a bit more to give the specific error and that is what
was running into the assert. Actually importing the document
itself does not hit the assert, nor does round-tripping it.

Since this is a stupid enough situation, and because I
can't think of a good way to word a SAL_WARN, and because
I can't say that I have detailed knowledge about this area,
I'll just remove the whole thing instead of leaving a warning.

Change-Id: If50b06f0103adca348ca314ebc92e7e4045967dc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/103133
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Caolán McNamara <caolanm@redhat.com>