Correctly read PNG into bitmaps N32BitTcA... formats (where alpha comes first)
This appears to be a regression introduced with
1ec439633003c06787774624dfc97cae44c0eb4f "Support for native 32bit Bitmap in VCL
and SVP (cairo) backend".  It caused CppunitTest_vcl_png_test to fail on
(big-endian) Linux s390x with

> vcl/qa/cppunit/png/PngFilterTest.cxx:176:PngFilterTest::testPng
> equality assertion failed
> - Expected: c[ff000040]
> - Actual  : c[0000ff40]

where eFormat happens to be ScanlineFormat::N32BitTcArgb, vs.
ScanlineFormat::N32BitTcBgra on e.g. Linux x86-64 (and which thus didn't notice
the lack of support for N32BitTcA... formats where alpha goes first instead of
last).

Change-Id: Id6030468718f6ef831b42f2b5ad7ba2c4c46a805
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/103240
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>