Restore the fix for tdf#126830
It was lost with 25936bdb9de3cd8791d171698316881ffe051395
("drop gtk2 support"), which was based on an older revision
of gtksalmenu.cxx.

In the meantime, the UI for settings icons on was hidden
(see tdf#123265), but this can't be an excuse for carrying
the broken code. The setting is also still available as an
expert config.

Change-Id: Iffc6342bb312230646399f2f85ef0211315f6c8a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/103644
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>