nss: restore manual pre-dependency build
We had some seldom build failures on Windows, with errors like:
PermissionError: [WinError 32] The process cannot access the file
because it is being used by another process: '..../nssckmdt.h'.

This happens, because of the "." / parent shell hack. Thinking
about it again, it doesn't prevent the parent make to run in
parallel to the "." directory make. So I tried to use a terminal
match-all rule like

ifneq (,$(filter .,$(DIRS)))
%::
	# empty terminal rule triggered
$(error can't happen)
endif

to stop the original parent make, but that doesn't work and the
$(error ...) is triggered.

So AFAIK I'm out of options here and have to restore the old
manual pre-dependency build variant - still much better then no
parallel build.

An alternative idea was to put the rest of the rules.mk in the
"else" of the terminal rule, to skip all normal rules, but this
still leaves out all rules from the rest of the make-files,
which might result in some hard to debug errors.

This reverts my upstream patch 15608:744881490c78.

Change-Id: I9e2e9e1ec9f35697c7853c92f60434f514cba5ab
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/103777
Tested-by: Jenkins
Reviewed-by: Tor Lillqvist <tml@collabora.com>
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>