Don't forward JDK home option to the cross-build
The configure flags on the command line always specify the host
config, so it doesn't make any sense to forward this setting in
case of a cross build. You can add a --with-jdk-home=... to the
--with-build-platform-configure-options flags, if the build
config finds the wrong build JDK. Also add this info to the
--with-jdk-home help.

For convenience (and because I don't want to adapt tinderboxes
again), this forwards the host JDK to the build for Android. And
explicitly disables Java for iOS build tools, as it used to be.

The whole --with-java handling should probably be removed in
favour of the --with-jdk-home. Do we have a real use case, where
we just need the Java interpreter but not the JDK (ok - the
document validators come to mind, but then you could also just
simply provide a JDK...)?

Change-Id: I328a15c090001bb025f4ba66d13b0cd020991fa9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/103878
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>