add EditViewCallbacks handler to support updating scrollbar state
the problem is that since...

commit 70a5c0eff6c3370351c22799ed1e68349dabe9fa
Date:   Tue Nov 22 16:21:20 2011 +0000

    tweak experimental gsoc multiline input bar, better resizing, enable scroll

that uses SetNofifyHdl to try and keep its scrollbar up to date, but that
SetNotifyHdl is also used by a11y to listen to the editengine and only one can
be set at a time, so with a11y enabled (the gtk default case) either a11y works
or the multiline scroll doesn't or vice versa.

Seeing as the a11y case is the very complicated case, leave a11y alone and
plot a route to disentangle the straightforward calc multiline edit from a11y.

Change-Id: Iedc7ffc39940354e8a05c0620944f617eee6b550
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/104080
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>