Resolves: tdf#137248 Fix wrong condition breaking non-group interpret
Regression from

    commit f84d3b1e57cede0a2b95576d5f59fd03215162c1
    CommitDate: Thu Sep 10 11:15:47 2020 +0200

        detect if a cell still needs interpreting after Interpret()

The bool return from ScFormulaCell::Interpret() does not indicate
if it was succesful but whether a group interpret was done. As
both calls here happen in a non-group context bailing out on that
if false (no group interpret) is wrong. Instead, ask the cell if
it still needs to be interpreted after having been interpreted.

Change-Id: I40eeeef65e1da3d729cb3fef550620b1ea0a5741
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/104261
Tested-by: Jenkins
Reviewed-by: Eike Rathke <erack@redhat.com>