sd: custom annotation marker and draw polygon annotation from PDF
This implements custom annotation marker, which overrides the
default marker drawing of comments in Draw/Impress. The
polygon is transported from the import of PDF to SdPdfFilter,
then the sd::Annotation is set-up, by settuing up the new class
CustomAnnotationMarker.

CustomAnnotationMarker also supports setting of the line color,
line width and fill color.

The OverlayPolyPolygon is the new class that is responsible for
the marker overlay, mainly to create the Primitive2D of the
marker, that will be shown on the screen.

This only implements Polygon PDF annotation sub-type.

Change-Id: Ic663c31c5b3db5c13179dde63c1a0b81159f4b80
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/104365
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>