Fix check for inclusion of <sys/sysmacros.h>
480ec88832d218dc144023c3bd3dde9dc590dc52 "Do not include <sys/sysmacros.h>
header unconditionally, but perform a check for it in the configure script" had
added an AC_DEFINE without a corresponding mention in any
config_host/config_*.h.in (we use AC_CONFIG_HEADERS), so the #ifdef was always
false.

The #include <sys/sysmacros.h> had been added with
eb29afe612c5015627c0197dd1b1c1ecb001e317 "WaE: include needed header" for the
declaration of a major function, but which is only used in #ifdef LINUX code
anyway.

Change-Id: I81b574c4a3e5fa2ef4e64a507b4841044217409b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/104351
Tested-by: Tor Lillqvist <tml@collabora.com>
Tested-by: Jenkins
Reviewed-by: Tor Lillqvist <tml@collabora.com>