tdf#136924 draw lines using field text color
Interestingly, these are not icon images, as I expected, but it's
a custom DrawingArea widget, which is completely drawn by some
algorithm.

Currently it uses the themes field color as background fill color
and an independent font color for the lines. So just use the
themed, matching field text color instead.

Regressed-by: 39e027c669fc0a91e223d7635c10adb5778586c7
Change-Id: I9f0073f41127207d2391ee23c9553512ffe4ba01
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/104422
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>
Tested-by: Jan-Marek Glogowski <glogow@fbihome.de>