remove ui for 'forceskia' AKA 'ignore skia denylist' (tdf#137159)
I originally copy&pasted this from the OpenGL code, but now that I
think of it, having an easy checkbox to make LO use drivers that LO
decides are faulty is a bad idea. There's still the expert
configuration if somebody insists (and if they're not expert
enough to find the expert setting then they better not mess with it),
but this is asking for unnecessary trouble.
This also solves the problem that the UI option is misleading. As
the description in Common.xcs says, "This one forces the use of Skia
even if the denylist would block the driver.", so the option is
independent of the 'enable skia' option, but the UI didn't make it
appear so.

Change-Id: I74bf3574f16899405272dbb3aec54580a5e3f0bb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/104425
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>