remove unnecessary `if (!true) return;` statement from Desktop::Init()
Seems that this specific piece of code was introduced in
commit 8c9e98212b97af0e38a1fc8ab9819463d4198a98 by Noel Grandin,
where `if (true)` wrapper was removed and `if (!true) return;`
seems to have been left over.

But the original `if (true)` statement was introduced in commit
bca00463739db0640adfba2bc39c1679a0036166 by Michael Meeks, where
`if ( m_aBootstrapError == BE_OK )` before starting IPC thread
was replaced with `if (true)`. I asked on IRC and they were not
sure for the reason of this change, so I will add Michael Meeks
as a reviewer of this patch, just in case.

Change-Id: I1ffb781464fa2fde152eb21bc61e4b244e8812d3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/104530
Reviewed-by: Michael Meeks <michael.meeks@collabora.com>
Tested-by: Jenkins