tdf#137660 send some focus event for push buttons
With KDE breeze we don't get any focus indicator for the keyboard
navigation in the start center, with either the kf5 or the gtk3
plugin using breeze. gtk3 with adwaita works fine, because it can
handle the ControlPart::Focus of the push button.

This behaviour was introduced in #i110915#. In the gen plugin
you get a small focus rect inside the button, but kf5 and gtk3
set mbNoFocusRectsForFlatButtons, which was introduced by commit
6c732a10867e28ea61ca2f6fe8b58887cd01ca3c, so there isn't any
indicator left for kf5 for the keyboard focus.

So this sends focus updates for the flat buttons, if the focus
rects are disabled and ControlPart::Focus is not supported, so
kf5 can render a focus indicator.

Change-Id: I1b587a7f9ec55b44369c4c8bc21039fe90aab9ea
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/104582
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>