Properly (un)tick Java checkbox in Java options in direct mode
While the Java settings cannot be changed when the direct mode
of the Java framework is used
(s. Change-ID Ife017f9b5c6c6488f84201dd78b23305c67bec1b,
"Gray out Java options when framework's direct mode is used"),
it still makes sense to (un)tick the (grayed out) checkbox which
indicates whether a JRE is used or not, which depends on whether
the JRE that is set (e.g. via env variable
'UNO_JAVA_JFW_JREHOME') is usable or not.

Change-Id: I13822bb7535e3f05cbc6ef83b3c82e2739c45ad6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/104064
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>