Don't make list of JREs active when parent frame is inactive
When the Java framework's direct mode is used,
there is no use in allowing the user to change the Java
options in the "Advanced" options page, so the corresponding
frame is grayed out, s. commit
45ab30104504062ac4e517a8401c9032273c6400
("Gray out Java options when framework's direct mode is used",
2020-10-21).

For the native gtk3 implementation, this also means that the
child widget holding the list of JREs ('m_xJavaList')
automatically remains grayed out as long as the parent widget
('m_xJavaFrame') is.
However, it turns out that this is not true for the
plain VCL implementation (used e.g. by the gen or kf5 VCL
plugins) where the child widget can be
made sensitive while the parent widget is insensitive.

Therefore, commit d6fc2f5bf91158d8c7d8a06c8a86808770726990
("Properly (un)tick Java checkbox in Java options in direct
mode", 2020-10-21) resulted in the widget holding the
list of JREs becoming active again when the
'EnableHdl_Impl' handler was called in the non-gtk3 case
(while the parent widget and the buttons to add JREs, edit
Java parameters and classpath would remained grayed out).

Make sure the widget holding the list of JREs remains
grayed out along with the whole frame holding the Java
options.

Change-Id: Ic46bf317afec9bc566493ec56ab5319a78050da0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/104657
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>