tdf#131679 sw: fix crash when copying fly via context menu
sw::DocumentContentOperationsManager::CopyImplImpl() is called with a
rPam that's on an SwOLENode.

The problem (which i can't reproduce in --enable-dbgutil build,
presumably for timing reasons) is that after the context menu pops up,
some idle layout runs and reformats the document and deletes a
SwFlyFrame and that calls SdrMarkView::UnmarkAllObj().

Then when SwFEShell::Copy() is called, it finds IsFrameSelected()
returns false, and it tries to copy normal text when the cursor is on
an SwOLENode.

Fix this in SwFlyFrame::FinitDrawObj() by first moving the cursor out
of any selected flys.

(regression from 74cdd53b4d01e45129d3115ef9439d99cf514df0
 - previously CopyImplImpl() would return early)

Change-Id: Id135fcc002c03c07c34fbdc0355f2895d8b6565b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/104697
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>