tdf#136555 apply window bg color for button boxes
The default background color of the frame is APPBACKGROUND, which
is matched to GetWorkspaceColor(). That color is visible in Writer
as the default document workspace background.

Originally I just saw the dark background of the button box in the
start center, and thought the workspace color in Qt is wrong and
changed it in the reverted commit, missing the change of the color
around the Writer document.

Now that the button boxes's background color isn't fixed via the
"branding" setting, we also must react to changed style settings,
so we override the DataChanged function.

This also reverts commit 56a45dc65876956369b37bf806b415d6c956e966.

Regressed-by: ffd166127679f89385c4e66760a7f4e5260136b0
Change-Id: I62396355054523baef1197a8a8af61c2f0d29ef3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/104740
Tested-by: Jenkins
Reviewed-by: Thorsten Wagner <thorsten.wagner.4@gmail.com>
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>