Add another return value to the PathFormat function, for WSL
This is just one step towards the goal of being able to build for
Windows on WSL (instead of Cygwin).

Unlike Cygwin, WSL can not use Windows pathnames at all. So to be able
to build for Windows on WSL, we will need to be much more careful in
keeping track of which pathnames are in Windows format and which are
in Unix format. Add a second return value for PathFormat that on WSL
is in Unix format.

Like in 2dde2f4865d71a255695edfee461a920c1a1f3b9, also add a simple
"unit test" (just for visual inspection) of the new functionality.

Change-Id: Ief4584243227677d0705de2c83e2a0bf54eac0a3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/104761
Tested-by: Jenkins
Reviewed-by: Tor Lillqvist <tml@collabora.com>