tdf#115588 write autoDataStyles for master-slides
If we don't trigger this, then the numeric styles won't be
written to the auto-styles and the style.xml will refer to
styles, that don't exists. In particular a reference to "N2",
where "N2" number-style wouldn't be defined.

Flat ODS, that writes all into one document, was unaffected by
this (because we trigger the autoDataStyles funtion in when also
exporting the content.

Change-Id: Iddfc264756d828b098ca8a351299c961b7e63dab
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/104760
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>