tdf#137413 EMF import: fix transparency in the PDF fallback case
Commit b35fa1d119a704697f2af730a78e5e45c49b10bc (tdf#136836 emfio: speed
up import of EMF import when the orig PDF is available, 2020-09-17)
improved both performance and correctness of the EMF import, in case it
had a PDF fallback.

It turns out that PDF fallback can be nominally non-transparent, and
still the EMF equivalent supports transparency.

Fix the problem by enabling transparency in the PDF-in-EMF case.

Change-Id: I4d1585a5db6f28bd9c9cb380b5f193f4d5edcc8d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/104849
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins