Tentative fix for crash in the iOS app
See https://github.com/CollaboraOnline/online/issues/403 .

It seems that perhaps we can just bypass the whole ImplHandlePaintHdl
handler function on iOS. Not sure whether this has any unintended
side-effects. Testing will show. (Sure, if this is the right fix, it
would make more sense to not create that idle paint thing at all. Feel
free.)

As discussed in the issue, there is a problem on Linux, too, that
might have the same root cause, that can be reproduced in a simple
"make run" scenario. It is possible that to fix that, the
ImplHandlePaintHdl function should simply return right away if
comphelper::LibreOfficeKit::isActive() is true. (And if that is done,
the #ifndef added here can be dropped.) But I am even less convinced
that such a change doesn't have any ill side-effects, and the symptom
is perhaps less serious than on iOS, so I will leave that to others to
investigate, for now.

Change-Id: Ie4c1c70c65746961fa0730cae348ecc9bcdccf1d
Signed-off-by: Tor Lillqvist <tml@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/104918
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/106257