Pass ExternalInfo up to caller from lcl_ScAddress_Parse_OOo()
... if external and sheet name was detected but a reference not
completely parsed, which may indicate an external sheet-local
name. This then in ScCompiler::IsSingleReference() is used to
reset mnCurrentSheetEndPos and mnCurrentSheetTab, which without
before was not done and may have lead to the external sheet-local
name being taken as a sheet-local name in the current document if
there was an identical one. External sheet-local names so far are
not supported at all.

The ExternalInfo is filled only if pSheetEndPos was passed but
then could be used by the caller to obtain a name later.
ExternalInfo::maTabName is filled only if a sheet name separator
was encountered.

Change-Id: I2414655a90a961af0e8ac2367cde7bdc4dcbb4c6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/105089
Tested-by: Jenkins
Reviewed-by: Eike Rathke <erack@redhat.com>