tdf#132536 free the cached VirtualDevices
Once again I forgot, that VclPtr is not really a shared pointer,
as someone has to release the contained instance. It just prevents
others from working on a dangling pointer and also freeing the
contained instance multiple times. Using a "Scoped" variant
in the vector instead of the loop doesn't compile.

This fixes the major memory leak when deleting the last
FontNameBox, which freed the caching VclPtr vector but not the
contained VirtualDevice instances. It seems there is still some
other minor leak left to fix, probably unrelated to the regressing
patch.

Found, after I realized the memory leak just happened with font
preview enabled...

Regressed-by: 45d27a1a3537494a0c25b23451e98c41654ce5c0
Change-Id: Ic0c60cd31bb6dabf03b0f8bd232390e784421588
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/105017
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>