tdf#128877 DOCX import: fix lost position of image cropped to shape
Regression from commit da9dedc13d197fd0951fe9151a340a5d570c04c6 (ooxml
import: supprt cropping to shape, 2019-05-13), which changed the type of
a cropped-to-shape image from drawing.GraphicObjectShape to
drawing.CustomShape.

drawing.GraphicObjectShape worked because GraphicImport::lcl_attribute()
in writerfilter/ had a check for drawing.GraphicObjectShape and did an
explicit setPosition().

Doing the same for bitmap-filled custom shapes would be an option, but
it would be ugly: scaling/translation/rotation/mirroring can only work
together if they are only applied once, and that should happen in oox/,
that's why we already have a mechanism to send the position from
writerfilter/ to oox/ for WPS shapes. (<a:xfrm> contains the size, but
not the position of the shape, so oox/ in itself could not know the
position.)

Fix the problem by improving ShapeContextHandler instead the pass the
position from writerfilter/ to oox/ for <pic:pic> as well, the same is
done for <wps:wsp> already since commit
6f8c8c6c940ce378089058731f6b94b1f88fc043 (DOCX drawingML shape import:
fix position when CustomShapeGeometry is set, 2013-11-21).

Change-Id: I74a60136d0ca8383e58948711b47858823f42437
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/105263
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins