pdf: use a common call to parse a object
Previously, only Lookup method called parsing correctly, taking
into account that the object is in its own stream and has its
own elements vector. Many other methods called parse with the
wrong - document element vector. This changes the code so that
a common method is called in all instances with the correct
elements vector as the input parameter.

Change-Id: I7092f7ce683f07065da15cfa548b06c019efeed4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/105491
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>